Extract cluster name
